% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctionsForecastData.R
\name{ConvertKelvinIntoCelsius}
\alias{ConvertKelvinIntoCelsius}
\title{Converts Kelvin into Celsius for a predefined set of variables}
\usage{
ConvertKelvinIntoCelsius(DWDForecastDf)
}
\arguments{
\item{DWDDataFrame}{data.frame. Extracted parameters}
}
\value{

}
\description{
if set to TRUE variables `TX`, `TTT`, `TN`, `T5cm` will be converted into Celsius.
}
\details{
Conversion factor Kelvin in Celsius is -273.15
}
\examples{
# print internal data.frame
ForecastDWDStationE273
ConvertKelvinIntoCelsius(DWDForecastDf = ForecastDWDStationE273)
}
