% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctionsForecastData.R
\name{CreateURLQuery}
\alias{CreateURLQuery}
\title{Creates a URL to query the weather data based on the ID of the weather station}
\usage{
CreateURLQuery(DWDWeatherStationID)
}
\arguments{
\item{DWDWeatherStationID}{character. Weather station ID as character.}
}
\value{
data.frame. URL weather station with ID. ID of the weather station.
}
\description{
Generates a URL in the form "https://opendata.dwd.de/weather/local_forecasts/mos/MOSMIX_L/single_stations/01001/kml/MOSMIX_L_LATEST_01001.kmz"
}
\examples{
CreateURLQuery(DWDWeatherStationID = "E273")
CreateURLQuery(DWDWeatherStationID = "E355")
}
