% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctionsForecastData.R
\name{DWDForecasting}
\alias{DWDForecasting}
\title{Request DWD Forecast data along a set of coordinates}
\usage{
DWDForecasting(
  IDsDWDWeatherStations,
  ParametersToExtract = c("RR1c", "ww", "T5cm", "TTT", "TN", "TX", "FF", "SunD", "N"),
  ConvertKelvinToCelsius = TRUE,
  tmap = TRUE,
  tmapOutputDir = tempdir()
)
}
\arguments{
\item{IDsDWDWeatherStations}{character. Vector with IDs of DWD weather stations}

\item{ParametersToExtract}{character. Vektor mit zu extrahierenden Parametern aus den DWD `kml`-Dateien}

\item{ConvertKelvinToCelsius}{bool. Convert Kelvin into Celius. default: `TRUE`}

\item{tmap}{bool. Generate map (html) with weather stations with tmap? default: `FALSE`}

\item{tmapOutputDir.}{character. tmap output directory. default: `tempdir()`}
}
\value{
list. Meta data and data.frame with forecast data of the extracted weather stations.
}
\description{
Request DWD Forecast data along a set of coordinates
}
\details{
Available parameter - see also data object `MOSMIXLParameter`:

- RR1c: Gesamtniederschlag (1-stündig) konsistent mit ww - Einheit kg/m2 - Stündlich

- ww: Signifikantes Wetter - Einheit % - Stündlich

- T5cm: Temperatur am Erdboden in 5cm Hoehe - Einheit K - Stündlich

- TTT: 2m-Temperatur - Einheit K - Stündlich

- TN: Minimumtemperatur - letzte 12 Stunden - Einheit K - 2xTag: 06:00 AM, 18:00 PM

- TX: Maximumtemperatur - letzte 12 Stunden - Einheit K - 2xTag: 06:00 AM, 18:00 PM

- FF: Windgeschwindigkeit - Einheit m/s - Stündlich

- SundD: Sonnenscheindauer innerhalb der letzten 24 Stunden - Einheit s - Täglich

- N: Gesamtbedeckung - Einheit % - Stündlich
}
\examples{
# Example 1: Weather stations with different available parameters
IDsDWDWeatherStationsExample1 <- c("11120", "06753", "74783", "E273", "E355")
# Request weather data
ForecastingDWDStationsExample1 <- DWDForecasting(IDsDWDWeatherStations = IDsDWDWeatherStationsExample1)
# Get a glimpse of the data
head(ForecastingDWDStationsExample1)
# Number of forecast time points of each station
table(ForecastingDWDStationsExample1$StationsID)

# Example 2: Weather stations with same available parameters
# Request weather data
ForecastingDWDStationsExample2 <- DWDForecasting(IDsDWDWeatherStations = c("E273", "10224", "A762", "K2689",
                                                                           "E355",  "H542",  "10210", "A981",
                                                                           "E078",  "H330",  "H573",  "P232",
                                                                           "10400", "H744",  "10442", "E426"))
# Get a glimpse of the data
head(ForecastingDWDStationsExample2)
# Number of forecast time points of each station
table(ForecastingDWDStationsExample2$StationsID)


#Example 3 - Special case: only a single weather station
ForecastingDWDStationsExample3 <- DWDForecasting(IDsDWDWeatherStations = "E273")
}
