% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctionsForecastData.R
\name{DWDForecastPackageHandling}
\alias{DWDForecastPackageHandling}
\title{Installation or loading of the required packages for obtaining the forecasting data of the DWD}
\usage{
DWDForecastPackageHandling()
}
\value{
message. Returns a message that the packages are loaded or not loaded. If packages are not loaded,
check carefully what happend.
}
\description{
Installation or loading of the required packages for obtaining the forecasting data of the DWD
}
\examples{
DWDForecastPackageHandling()
}
