% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctionsForecastData.R
\name{ExtractAvailableVariables}
\alias{ExtractAvailableVariables}
\title{Extract available parameters of a weather station - returns either a list or a vector with names}
\usage{
ExtractAvailableVariables(
  DWDXMLData,
  ParametersToExtract = c("RR1c", "ww", "T5cm", "TTT", "TN", "TX", "FF", "SunD", "N"),
  DataFrame = TRUE
)
}
\arguments{
\item{DWDXMLData}{XMLInternalDocument / XMLAbstractDocument. Extracted XML document from DWD kmz file.}

\item{ParametersToExtract}{character vector. Vector with variable names to extract.
default: `c("RR1c", "ww", "T5cm", "TTT", "TN", "TX", "FF","SunD", "N")`}

\item{DataFrame}{bool. Return a data.frame or a character vector with the variable names.
default `TRUE`  returns a data.frame.}
}
\value{
either character vector or data.frame. Returns either a data.frame with parameters or a data.frame with the
extracted data.
}
\description{
Extract available parameters of a weather station - returns either a list or a vector with names
}
\examples{
## Station K057
TestRequestData <-
  data.frame(DWDWeatherStationID = "K057",
             URL = "https://opendata.dwd.de/weather/local_forecasts/mos/MOSMIX_L/single_stations/K057/kml/MOSMIX_L_LATEST_K057.kmz")
K057PathAndFolder <- RequestMOSMIXLData(DWDURLsDF = TestRequestData,
                                        ColDWDStationID = "DWDWeatherStationID",
                                        ColURL = "URL")
# Extract from XML file
K057ExtractedForecastData <- ExtractDWDForecastingData(PathFolderFileName = K057PathAndFolder)
# Extract station data
ExtractXMLStationData(DWDXMLData = K057ExtractedForecastData)
# Extract names of available parameters of a weather station
ExtractAvailableVariables(DWDXMLData = K057ExtractedForecastData, DataFrame = FALSE)
# data.frame
ExtractAvailableVariables(DWDXMLData = K057ExtractedForecastData, DataFrame = TRUE)
}
