% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctionsForecastData.R
\name{ExtractDWDForecastingData}
\alias{ExtractDWDForecastingData}
\title{DWD Forecasting file - kmz format - unpack and return as XML document}
\usage{
ExtractDWDForecastingData(PathFolderFileName)
}
\arguments{
\item{PfadZumOrdnerUndDateiName}{character. Folder where the kmz file is stored}
}
\value{
XMLInternalDocument / XMLAbstractDocument. Generated XML document for further processing
}
\description{
DWD Forecasting file - kmz format - unpack and return as XML document
}
\examples{
## Station K057
TestRequestData <-
  data.frame(DWDWeatherStationID = "K057",
             URL = "https://opendata.dwd.de/weather/local_forecasts/mos/MOSMIX_L/single_stations/K057/kml/MOSMIX_L_LATEST_K057.kmz")
K057PathAndFolder <- RequestMOSMIXLData(DWDURLsDF = TestRequestData,
                                        ColDWDStationID = "DWDWeatherStationID",
                                        ColURL = "URL")
# Extract from XML file
K057ExtractedForecastData <- ExtractDWDForecastingData(PathFolderFileName = K057PathAndFolder)
class(K057ExtractedForecastData)
}
