% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ForecastDWDStationE273}
\alias{ForecastDWDStationE273}
\title{5 rows of data from DWD station E273}
\format{
A data.frame with 5 rows and 11 variables:
\describe{
  \item{datetime}{chr, date/time variable saved as character}
  \item{TX}{num, eine kurze Informazion des Projektes: enthält Vertragsnummer, Golfplatzort, Vertragsbeginn}
  \item{TTT}{num, Die geographische Breite eines Golfplatzes}
  \item{TN}{num, Die geographische Länge eines Golfplatzes}
  \item{T5cm}{num, Löcheranzahl je Golfplatz}
  \item{FF}{num, eindeutige DWD Stationsnummer}
  \item{N}{num, Entfernung vom Golfplatz zur nächsten DWD Station, km}
  \item{ww}{num, Entfernung vom Golfplatz zur nächsten DWD Station, km}
  \item{RR1c}{num, Entfernung vom Golfplatz zur nächsten DWD Station, km}
  \item{SunD}{num, Entfernung vom Golfplatz zur nächsten DWD Station, km}
  \item{StationsID}{chr, Entfernung vom Golfplatz zur nächsten DWD Station, km}


}
}
\source{
{"data/"}
}
\usage{
ForecastDWDStationE273
}
\description{
The dataset contains 5 lines of a typical DWD weather station output generated with the functions in the
package `DWDForecasting`
}
\keyword{datasets}
