% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOSMIXLParameter}
\alias{MOSMIXLParameter}
\title{MOSMIX-L parameter description}
\format{
A data.frame with 9 rows and 4 variables:
\describe{
  \item{abbreviation}{chr, variablename}
  \item{description}{chr, short description}
  \item{description_en}{chr, short description in English}
  \item{unit}{chr, please note: unit can be changed during processing, e.g. K in °C, seconds in minutes}
  \item{frequency}{chr, measured how often?}
}
}
\source{
{"data/"}
}
\usage{
MOSMIXLParameter
}
\description{
The dataset contains the Deutsche Wetterdienst (DWD) MOSMIX-L parameter description in German language
}
\keyword{datasets}
