% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctionsForecastData.R
\name{RequestMOSMIXLData}
\alias{RequestMOSMIXLData}
\title{Download the forecast weather data of the DWD}
\usage{
RequestMOSMIXLData(
  DWDURLsDF,
  ColDWDStationID = "DWDWeatherStationID",
  ColURL = "URL"
)
}
\arguments{
\item{DWDURLsDF}{data.frame. URL weather station with ID. ID of the weather station}

\item{ColDWDStationID}{character. Weatherstation DWD ID}

\item{ColURL}{character. Column name generated query URL}
}
\value{
character. Returns the path and folder of the download.
}
\description{
Downloads the latest L-type DWD MOSMIX forecast file: MOSMIX-L: produced 4 times a day up to 240 hours
ahead, provides up to 115 parameters (location dependent)
}
\details{
Stored in the tempdir(). Please note, however, that this does not save the files permanently,
since the tempdir is deleted again after an R session is ended.
}
\examples{
TestRequestData <-
  data.frame(DWDWeatherStationID = "11120",
             URL = "https://opendata.dwd.de/weather/local_forecasts/mos/MOSMIX_L/single_stations/11120/kml/MOSMIX_L_LATEST_11120.kmz")
RequestMOSMIXLData(DWDURLsDF = TestRequestData, ColDWDStationID = "DWDWeatherStationID", ColURL = "URL")
}
